Migrate confidence threshold service to float for Stereo ace BslDepthMinConf
setConfidenceThreshold now takes const double& and its ROS service switches from SetIntegerValue to SetFloatValue, mirroring the depth min/max float migration. Adds a Stereo ace override writing/reading BslDepthMinConf (CFloat, normalized [0,1]); blaze keeps its integer ConfidenceThreshold node via std::llround. Hardware-validated on my_stereo_ace (X-100-547C): 0.5 -> 0.75 readback -> restored 0.5, service returns success/done.

+++ pylon_ros2_camera_component/include/internal/impl/pylon_ros2_camera_base.hpp
@@ -4996,7 +4996,7 @@ std::string PylonROS2CameraImpl<CameraTraitT>::setAmbiguityFilterThreshold(const
 }
 
 template <typename CameraTraitT>
-std::string PylonROS2CameraImpl<CameraTraitT>::setConfidenceThreshold(const int& threshold __attribute__((unused)))
+std::string PylonROS2CameraImpl<CameraTraitT>::setConfidenceThreshold(const double& threshold __attribute__((unused)))
 {
     RCLCPP_DEBUG(LOGGER_BASE, "Feature not available for this camera type");
     return "Feature not available for this camera type";
